library now uses 16-bit indices

+++ clr/DNP3CLRAdapter/MasterDataObserverAdapter.cpp
@@ -13,27 +13,27 @@ void MasterDataObserverAdapter::Start()
 	proxy->Start();
 }
 
-void MasterDataObserverAdapter::Update(const opendnp3::Binary& arPoint, size_t aIndex)
+void MasterDataObserverAdapter::Update(const opendnp3::Binary& arPoint, uint16_t aIndex)
 {
 	proxy->Update(Conversions::convertMeas(arPoint), aIndex);
 }
 
-void MasterDataObserverAdapter::Update(const opendnp3::Analog& arPoint, size_t aIndex)
+void MasterDataObserverAdapter::Update(const opendnp3::Analog& arPoint, uint16_t aIndex)
 {
 	proxy->Update(Conversions::convertMeas(arPoint), aIndex);
 }
 
-void MasterDataObserverAdapter::Update(const opendnp3::Counter& arPoint, size_t aIndex)
+void MasterDataObserverAdapter::Update(const opendnp3::Counter& arPoint, uint16_t aIndex)
 {
 	proxy->Update(Conversions::convertMeas(arPoint), aIndex);
 }
 
-void MasterDataObserverAdapter::Update(const opendnp3::ControlStatus& arPoint, size_t aIndex)
+void MasterDataObserverAdapter::Update(const opendnp3::ControlStatus& arPoint, uint16_t aIndex)
 {
 	proxy->Update(Conversions::convertMeas(arPoint), aIndex);
 }
 
-void MasterDataObserverAdapter::Update(const opendnp3::SetpointStatus& arPoint, size_t aIndex)
+void MasterDataObserverAdapter::Update(const opendnp3::SetpointStatus& arPoint, uint16_t aIndex)
 {
 	proxy->Update(Conversions::convertMeas(arPoint), aIndex);
 }
